% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names}
\alias{names}
\title{Names of some characters in The Man in the High Castle}
\format{A vector with 23 rows.
\describe{
  \item{name}{name, with prefixes}
}}
\source{
Online search
}
\usage{
names
}
\description{
A vector containing names of 23 characters in The Man in the High Castle.
}
\keyword{datasets}
