% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names1}
\alias{names1}
\title{Names of some judges in the US Supreme Court and the UK Court of Appeal}
\format{A vector with 17 rows.
\describe{
  \item{name}{name, with prefixes}
}}
\source{
Online search
}
\usage{
names1
}
\description{
A vector containing names of 17 judges across the two courts
}
\keyword{datasets}
