% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names2}
\alias{names2}
\title{Names of some companies in the US}
\format{A vector with 16 rows.
\describe{
  \item{name}{name, with prefixes or suffixes}
}}
\source{
Online search
}
\usage{
names2
}
\description{
A vector containing names of 16 US companies
}
\keyword{datasets}
