% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_name.R
\name{preproc_name}
\alias{preproc_name}
\title{Pre-processes names for further analysis}
\usage{
preproc_name(names, stopwords = c("^MRS ", "^MR ", "^MISS ", "^DR ",
  "^MS ", " JR$", " SR$", " MD$"))
}
\arguments{
\item{names}{A string or character vector of names.}

\item{stopwords}{A vector of stopwords to remove from the names. Takes in Regex as well.}
}
\value{
The character vector with the stopwords in \code{stopwords} removed.
}
\description{
Pre-processes names for further analysis
}
\examples{
persons <- c("Mr. John Smith", "Mrs. Helen Smith", "Ms. Juliana Crain")
stopwords1 <- c("Mr", "Mrs", "Ms")
preproc_name(persons, stopwords=stopwords1)
}
